introduce util for e2e tests de-duplicate code
